moved all graphql files together under 'lib/graphql';
